@@ -25,29 +25,29 @@
   },
   "homepage": "https://github.com/jamesgower/typescript-react-boilerplate#readme",
   "devDependencies": {
-    "@babel/cli": "^7.5.5",
-    "@babel/core": "^7.5.5",
+    "@babel/cli": "^7.15.7",
+    "@babel/core": "^7.15.5",
     "@babel/plugin-proposal-class-properties": "^7.5.5",
     "@babel/plugin-proposal-numeric-separator": "^7.2.0",
     "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.3.3",
+    "@babel/preset-env": "^7.15.6",
+    "@babel/preset-react": "^7.14.5",
+    "@babel/preset-typescript": "^7.15.0",
     "@types/react": "^16.8.24",
     "@typescript-eslint/eslint-plugin": "^1.13.0",
     "@typescript-eslint/parser": "^1.13.0",
     "babel-loader": "^8.0.6",
     "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^3.1.0",
-    "eslint": "^5.3.0",
-    "eslint-config-airbnb": "^17.1.1",
-    "eslint-config-airbnb-typescript": "^4.0.1",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb": "^18.2.1",
+    "eslint-config-airbnb-typescript": "^14.0.0",
     "eslint-config-prettier": "^6.0.0",
-    "eslint-config-react": "^1.1.7",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-react": "^7.14.3",
+    "eslint-plugin-react": "^7.26.0",
+    "eslint-plugin-react-redux": "^3.3.2",
     "file-loader": "^4.1.0",
     "html-webpack-plugin": "^3.2.0",
     "mini-css-extract-plugin": "^0.8.0",
@@ -61,12 +61,17 @@
     "stylelint-prettier": "^1.1.1",
     "stylelint-scss": "^3.9.2",
     "terser-webpack-plugin": "^1.4.1",
-    "typescript": "^3.5.3",
+    "typescript": "^4.4.3",
     "webpack": "^4.39.0",
     "webpack-cli": "^3.3.6",
     "webpack-dev-server": "^3.7.2"
   },
   "dependencies": {
+    "@babel/plugin-proposal-optional-chaining": "^7.14.5",
+    "@material-ui/core": "^4.12.3",
+    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@types/express": "^4.17.13",
+    "@types/node": "^16.9.4",
     "@types/react-dom": "^16.8.5",
     "@types/react-router-dom": "^4.3.4",
     "bootstrap-css-only": "^4.3.1",
@@ -74,18 +79,19 @@
     "core-js": "^3.1.4",
     "dayjs": "^1.8.16",
     "express": "^4.17.1",
-    "node-sass": "^4.12.0",
+    "history": "^5.0.1",
+    "node-sass": "^4.14.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "react-redux": "^7.1.1",
+    "react-redux": "^7.2.5",
     "react-router-dom": "^5.0.1",
     "reactstrap": "^8.1.0",
     "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
     "regenerator-runtime": "^0.13.3",
-    "socket.io": "^2.3.0",
-    "socket.io-client": "^2.3.0",
-    "ts-node": "^8.4.1",
-    "uuid": "^3.3.3"
+    "socket.io": "^2.4.1",
+    "socket.io-client": "^2.4.0",
+    "ts-node": "^8.10.2",
+    "uuid": "^3.4.0"
   }
 }
