@@ -1,44 +1,32 @@
 import { Socket } from "socket.io";
 import { UserProps } from "../utils/interfaces/user.i";
 
-export interface ChatterState {
-  room?: RoomsState;
-  user?: UsersState;
+export interface AppState {
+  room: RoomsState;
+  user: UsersState;
 }
 
 export interface UsersState {
-  id?: string;
-  name?: string;
-  activeRoom?: string;
+  id: string;
+  name: string;
+  activeRoom: string;
 }
 
 export interface LandingState {
   name: string;
   room: string;
-  nameError: boolean;
-  roomError: boolean;
-  dropdownOpen: boolean;
+  nameError: string;
+  roomError: string;
   submitted: boolean;
-}
-
-export interface LandingDispatchProps {
-  setUser: (user) => void;
-  addRoom: (name) => void;
 }
 
 export interface RoomsState {
   socket?: Socket;
   currentRoom: string;
-  rooms: {
-    [x: string]: {
-      name: string;
-      activeUsers: number;
-      messages: Message[];
-    };
-  };
+  rooms: RoomProps[];
 }
 
-interface RoomProps {
+export interface RoomProps {
   name: string;
   messages: MessageProps[];
   activeUsers: number;
@@ -66,7 +54,7 @@ export interface ChatState {
 }
 
 export interface MessageProps {
-  type: string;
+  messageType: string;
   message: Message;
   admin?: boolean;
   location?: boolean;
@@ -77,18 +65,4 @@ export interface Message {
   createdAt: number;
   text?: string;
   url?: string;
-}
-
-export interface ChatDispatchProps {
-  room?: RoomsState;
-  user?: UserProps;
-  name?: string;
-  socket?: Socket;
-  messages?: Message[];
-  setSocket: (socket: Socket) => void;
-  setUser: (user: UserProps) => void;
-  addMessage: (message: Message) => void;
-  clearMessages: () => void;
-  changeRoom: (room) => void;
-  setActiveRoom: (activeRoom) => void;
 }
