@@ -1,4 +1,4 @@
-import { RoomsState } from "../interfaces/components.i";
+import { RoomsState, RoomProps, MessageProps } from "../interfaces/components.i";
 import {
   ChangeRoomAction,
   CHANGE_ROOM,
@@ -12,8 +12,9 @@ import {
   SetActiveRoomAction,
   ADD_ROOM,
   AddRoomAction,
+  SET_ACTIVE_USERS,
+  SetActiveUsersAction,
 } from "../interfaces/actions.i";
-import { getRoomVariable } from "../utils/room";
 
 type RoomActionsTypes =
   | ChangeRoomAction
@@ -21,80 +22,90 @@ type RoomActionsTypes =
   | AddMessageAction
   | ClearMessagesAction
   | SetActiveRoomAction
-  | AddRoomAction;
+  | AddRoomAction
+  | SetActiveUsersAction;
 
 const defaultRoomState: RoomsState = {
   socket: null,
-  rooms: {
-    react: {
+  rooms: [
+    {
       name: "React & Redux",
       messages: [],
       activeUsers: 0,
     },
-    es6: {
+    {
       name: "ES6+ JavaScript",
       messages: [],
       activeUsers: 0,
     },
-    frameworks: {
+    {
       name: "New Frameworks",
       messages: [],
       activeUsers: 0,
     },
-    discuss: {
+    {
       name: "Discuss",
       messages: [],
       activeUsers: 0,
     },
-    design: {
+    {
       name: "Design Ideas",
       messages: [],
       activeUsers: 0,
     },
-    chill: {
+    {
       name: "Chill Zone",
       messages: [],
       activeUsers: 0,
     },
-  },
+  ],
   currentRoom: null,
 };
 
 export default (state = defaultRoomState, action: RoomActionsTypes): RoomsState => {
+  const roomName = state.currentRoom;
+  // const currentRoom = state.rooms.find((room): boolean => room.name === roomName);
   switch (action.type) {
     case SET_ACTIVE_ROOM: {
       const { activeRoom } = action;
       return {
         ...state,
-        rooms: {
-          ...state.rooms,
-          [activeRoom]: {
-            ...state.rooms[activeRoom],
-            activeUsers: state.rooms[activeRoom].activeUsers + 1,
+        rooms: state.rooms.map(
+          (room): RoomProps => {
+            if (room.name === activeRoom) {
+              return {
+                ...room,
+                activeUsers: room.activeUsers + 1,
+              };
+            }
+            return room;
           },
-        },
+        ),
         currentRoom: activeRoom,
       };
     }
     case CHANGE_ROOM: {
-      const prevRoomName = state.currentRoom;
-      const prevRoom = state.rooms[prevRoomName];
-      const nextRoomName = getRoomVariable(action.newRoom);
-      const nextRoom = state.rooms[nextRoomName];
+      const nextRoom = state.rooms.find((room): boolean => room.name === action.newRoom);
       return {
         ...state,
-        rooms: {
-          ...state.rooms,
-          [prevRoomName]: {
-            ...prevRoom,
-            activeUsers: prevRoom.activeUsers - 1,
+        rooms: state.rooms.map(
+          (room): RoomProps => {
+            if (room.name === roomName) {
+              return {
+                ...room,
+                activeUsers: room.activeUsers - 1,
+              };
+            }
+            if (room.name === nextRoom.name) {
+              return {
+                ...room,
+                activeUsers: room.activeUsers + 1,
+              };
+            }
+            return room;
           },
-          [nextRoomName]: {
-            ...nextRoom,
-            activeUsers: nextRoom.activeUsers + 1,
-          },
-        },
-        currentRoom: getRoomVariable(action.newRoom),
+        ),
+        currentRoom: action.newRoom,
       };
     }
     case SET_SOCKET:
@@ -103,47 +114,79 @@ export default (state = defaultRoomState, action: RoomActionsTypes): RoomsState 
         socket: action.socket,
       };
     case ADD_MESSAGE: {
-      const currentRoomName = state.currentRoom;
-      const currentRoom = state.rooms[currentRoomName];
+      const roomName = state.currentRoom;
+      const currentRoom = state.rooms.find((room): boolean => room.name === roomName);
+      console.log(action, roomName);
+      const newMessage: MessageProps = {
+        type: action.type,
+        messageType: action.messageType,
+        admin: action.messageType === "message-admin",
+        location: action.location,
+        message: action.message,
+      };
       return {
         ...state,
-        rooms: {
-          ...state.rooms,
-          [currentRoomName]: {
-            ...currentRoom,
-            messages: [...currentRoom.messages, action.message],
+        rooms: state.rooms.map(
+          (room): RoomProps => {
+            if (room.name === roomName) {
+              return {
+                ...room,
+                messages: [...room.messages, newMessage],
+              };
+            }
+            return room;
           },
-        },
+        ),
       };
     }
     case CLEAR_MESSAGES: {
-      const currentRoomName = state.currentRoom;
-      const currentRoom = state.rooms[currentRoomName];
       return {
         ...state,
-        rooms: {
-          ...state.rooms,
-          [currentRoomName]: {
-            ...currentRoom,
-            messages: [],
+        rooms: state.rooms.map(
+          (room): RoomProps => {
+            if (room.name === roomName) {
+              return {
+                ...room,
+                messages: [],
+              };
+            }
+            return room;
           },
-        },
+        ),
+      };
+    }
+    case SET_ACTIVE_USERS: {
+      return {
+        ...state,
+        rooms: state.rooms.map(
+          (room): RoomProps => {
+            if (room.name === action.room) {
+              return {
+                ...room,
+                activeUsers: action.users,
+              };
+            }
+            return room;
+          },
+        ),
       };
     }
     case ADD_ROOM: {
-      const { name } = action;
-      const nameVar = name.split(" ")[0];
+      const newRoom: RoomProps = {
+        name: action.name,
+        activeUsers: 0,
+        messages: [],
+      };
+      if (state.rooms.findIndex((room): boolean => room.name === newRoom.name) === -1) {
+        return {
+          ...state,
+          rooms: [...state.rooms, newRoom],
+          currentRoom: newRoom.name,
+        };
+      }
       return {
         ...state,
-        rooms: {
-          ...state.rooms,
-          [nameVar]: {
-            name,
-            activeUsers: 0,
-            messages: [],
-          },
-        },
-        currentRoom: nameVar,
+        currentRoom: newRoom.name,
       };
     }
     default:
