@@ -3,5 +3,5 @@ import { UserProps } from "./user.i";
 export interface RoomProps {
   id: string;
   room: string;
-  activeUsers: UserProps[];
+  activeUsers: number;
 }
